# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Joe Kuntani Platform API
  version: 0.1.0
  description: Versioned contract for the standalone Joe Kuntani platform.
  license:
    name: Proprietary - all rights reserved
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Relative API base; clients configure the environment origin.
security: []
...
paths:
...
  /api/admin/auth/login:
    post:
      operationId: loginStaff
      summary: Start a staff session with password verification
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Password accepted; complete MFA when indicated
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /api/admin/auth/mfa/setup:
    get:
      operationId: getStaffMFASetup
      summary: Return authenticator enrollment material for a password-accepted, MFA-pending session
      tags: [Authentication]
      security: [{ staffSession: [] }]
      responses:
        "200":
          description: Enrollment secret and otpauth URI for the pending MFA challenge
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MFASetupResponse" }
        "401": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /api/admin/auth/mfa/verify:
    post:
      operationId: verifyStaffMFA
      summary: Verify TOTP and rotate the staff session token
      tags: [Authentication]
      security: [{ staffSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MFAVerificationRequest" }
      responses:
        "200":
          description: MFA verified and session rotated
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthenticatedResponse" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /api/admin/auth/me:
    get:
      operationId: getStaffSession
      summary: Read the authenticated staff principal
      tags: [Authentication]
      security: [{ staffSession: [] }]
      responses:
        "200":
          description: Authenticated principal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffPrincipal" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/admin/auth/logout:
    post:
      operationId: logoutStaff
      summary: Revoke the current staff session
      tags: [Authentication]
      security: [{ staffSession: [], csrfToken: [] }]
      responses:
        "204": { description: Session revoked and cookies expired }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/admin/auth/permissions:
    get:
      operationId: getStaffPermissionCatalog
      summary: Read the role and permission catalog
      tags: [Authentication]
      security: [{ staffSession: [] }]
      responses:
        "200":
          description: Role permission matrix
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [permissions, roles]
                properties:
                  permissions:
                    type: array
                    items: { type: string }
                  roles:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      required: [role, label, permissions]
                      properties:
                        role: { type: string }
                        label: { type: string }
                        permissions:
                          type: array
                          items: { type: string }
        "401": { $ref: "#/components/responses/Problem" }
  /api/admin/auth/me/profile:
    patch:
      operationId: updateStaffProfile
      summary: Update the signed-in staff display name
      tags: [Authentication]
      security: [{ staffSession: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
      responses:
        "204": { description: Profile updated }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/admin/auth/me/password:
    post:
      operationId: changeStaffPassword
      summary: Change the signed-in staff password
      tags: [Authentication]
      security: [{ staffSession: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [current_password, new_password]
              properties:
                current_password: { type: string, minLength: 1 }
                new_password: { type: string, minLength: 12 }
      responses:
        "204": { description: Password changed }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/admin/auth/me/preferences:
    get:
      operationId: getStaffPreferences
      summary: Read signed-in staff preferences
      tags: [Authentication]
      security: [{ staffSession: [] }]
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffPreferences" }
        "401": { $ref: "#/components/responses/Problem" }
    put:
      operationId: saveStaffPreferences
      summary: Save signed-in staff preferences
      tags: [Authentication]
      security: [{ staffSession: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffPreferences" }
      responses:
        "204": { description: Preferences saved }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/admin/auth/users:
    get:
      operationId: listStaffUsers
      summary: List staff users
      description: Requires users:manage (Administrator only).
      tags: [Authentication]
      security: [{ staffSession: [] }]
      responses:
        "200":
          description: Staff directory
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [users]
                properties:
                  users:
                    type: array
                    items: { $ref: "#/components/schemas/StaffUser" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
    post:
      operationId: provisionStaffUser
      summary: Provision a staff user
      description: Requires users:manage (Administrator only).
      tags: [Authentication]
      security: [{ staffSession: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, email, password, role]
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
                email: { type: string, format: email }
                password: { type: string, minLength: 12 }
                role:
                  type: string
                  enum:
                    [administrator, booking_manager, content_editor, analyst]
      responses:
        "201":
          description: Staff user created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [id]
                properties:
                  id: { type: string, format: uuid }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/admin/auth/users/{userID}/disable:
    post:
      operationId: disableStaffUser
      summary: Disable a staff user and atomically revoke every active session
      description: Requires the server-side users:manage permission (Administrator only).
      tags: [Authentication]
      security: [{ staffSession: [], csrfToken: [] }]
      parameters:
        - name: userID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: User disabled and sessions revoked }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/admin/auth/users/{userID}/role:
    patch:
      operationId: updateStaffRole
      summary: Update a staff user's role
      description: Requires users:manage (Administrator only). Revokes the target user's sessions.
      tags: [Authentication]
      security: [{ staffSession: [], csrfToken: [] }]
      parameters:
        - name: userID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role:
                  type: string
                  enum:
                    [administrator, booking_manager, content_editor, analyst]
      responses:
        "204": { description: Role updated }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    staffSession:
      type: apiKey
      in: cookie
      name: jk_admin_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  responses:
    Problem:
      description: Request rejected
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 12, maxLength: 1024 }
    LoginResponse:
      type: object
      additionalProperties: false
      required: [mfa_required]
      properties:
        mfa_required: { type: boolean }
    MFAVerificationRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code: { type: string, pattern: "^[0-9]{6}$" }
    MFASetupResponse:
      type: object
      additionalProperties: false
      required: [email, secret, otpauth_uri, issuer, digits, period_seconds]
      properties:
        email: { type: string, format: email }
        secret: { type: string, minLength: 16, maxLength: 128 }
        otpauth_uri: { type: string, format: uri }
        issuer: { type: string }
        digits: { type: integer, const: 6 }
        period_seconds: { type: integer, const: 30 }
    AuthenticatedResponse:
      type: object
      additionalProperties: false
      required: [authenticated]
      properties:
        authenticated: { type: boolean, const: true }
    StaffPrincipal:
      type: object
      additionalProperties: false
      required:
        [
          id,
          name,
          email,
          role,
          mfa_verified,
          mfa_enabled,
          permissions,
          preferences,
        ]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role:
          type: string
          enum: [administrator, booking_manager, content_editor, analyst]
        mfa_verified: { type: boolean }
        mfa_enabled: { type: boolean }
        permissions:
          type: array
          items: { type: string }
        preferences: { $ref: "#/components/schemas/StaffPreferences" }
    StaffPreferences:
      type: object
      additionalProperties: false
      required:
        [email_product_updates, email_security_alerts, dense_ui, timezone]
      properties:
        email_product_updates: { type: boolean }
        email_security_alerts: { type: boolean }
        dense_ui: { type: boolean }
        timezone: { type: string, minLength: 1, maxLength: 64 }
    StaffUser:
      type: object
      additionalProperties: false
      required: [id, name, email, role, status, mfa_enabled, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role:
          type: string
          enum: [administrator, booking_manager, content_editor, analyst]
        status: { type: string }
        mfa_enabled: { type: boolean }
        updated_at: { type: string, format: date-time }
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference